changed all software quantized modules
