+++ machop/chop/passes/graph/transforms/quantize/quantized_modules/__init__.py
@@ -59,7 +59,6 @@ from .relu import (
 )
 from.layer_norm import (
     LayerNormInteger,
-    LayerNormHWInteger
 )
 
 quantized_module_map = {
@@ -111,5 +110,5 @@ quantized_module_map = {
     "relu_block_fp": ReLUBlockFP,
     "relu_binary": ReLUBinary,
     "relu_ternary": ReLUTernary,
-    "layer_norm_integer": LayerNormHWInteger,
+    "layer_norm_integer": LayerNormInteger,
 }
